設定画面をサブコンポーネントへ分離
