+++ src/components/IndexPage/SettingsPageContent/LifeInput.tsx
