+++ src/components/IndexPage/SettingsPageContent/MaxLifeInput.tsx
