+++ src/components/IndexPage/SettingsPageContent/ResetSettingsButton.tsx
@@ -1,6 +1,6 @@
 import { Button } from "@headlessui/react";
 import React from "react";
-import { SavedDataManager } from "./utils";
+import { SavedDataManager } from "../utils";
 
 type Props = {
   clearSavedData: SavedDataManager["clearSavedData"];
